gui: move UI-specific argument functions to gui_args
